* 기능 추가 - 필기 시험 응시, 채점 등에 대한 UI 및 로직 구현

+++ app/src/main/java/com/example/wemakepass/network/api/ExamInfoAPI.java
@@ -30,5 +30,5 @@ public interface ExamInfoAPI {
     @LoginRequired
     @GET(BASE_URI + "subject_list")
     Observable<Response<List<String>>> subjectList(
-            @Query("examId") int examId);
+            @Query("examId") long examId);
 }
